% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\name{forestplotter}
\alias{forestplotter}
\alias{forestplotter.default}
\alias{forestplotter.tidy_subgroup_tbl}
\title{Method to create forest plot for \code{tidy_subgroup} tibble}
\usage{
forestplotter(x, ...)

\method{forestplotter}{default}(
  data,
  est,
  lower,
  upper,
  sizes = 0.4,
  ref_line = ifelse(x_trans \%in\% c("log", "log2", "log10"), 1, 0),
  vert_line = NULL,
  ci_column,
  is_summary = NULL,
  xlim = NULL,
  ticks_at = NULL,
  ticks_digits = NULL,
  ticks_minor = NULL,
  arrow_lab = NULL,
  x_trans = "none",
  xlab = NULL,
  footnote = NULL,
  title = NULL,
  nudge_y = 0,
  fn_ci = makeci,
  fn_summary = make_summary,
  index_args = NULL,
  theme = NULL,
  ...
)

\method{forestplotter}{tidy_subgroup_tbl}(
  x,
  terms = x$term[[2]],
  label = "Subgroup",
  .after_stats = c(`Est. (95\% CI)` =
    "{format.ci(estimate, conf.low, conf.high, digits=1)}", `p-value` =
    "{format.pval(p.value, digits=2, eps=1e-3)}"),
  .before_stats = NULL,
  .show_p.value = TRUE,
  ...
)
}
\arguments{
\item{x}{a tibble create by tidy_subgroup}

\item{...}{additional parameters passed to [forestploter::forest()].}

\item{terms}{<character> which term to show forest plot}

\item{label}{[\code{"Subgroup"}] A character specifying the label of the subgroup column}

\item{.before_stats}{[\code{NULL}] <NULL|character> stat name in x to forest plot that are shown before the CI}
}
\value{
a plot
}
\description{
This function add a method to create forest plot to tidied subgroup models.
}
\author{
Trinh Dong
}
