% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\name{format.ci}
\alias{format.ci}
\title{Small helper for formatting confidence interval}
\usage{
\method{format}{ci}(est, lower, upper, digits = 1)
}
\arguments{
\item{est}{point estimate}

\item{lower, upper}{the lower and upper bound of CI}

\item{digits}{number of digits}
}
\description{
This is a small helper to format the CI in foresplot
}
