% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ingroup.R
\name{tidy_withingroup}
\alias{tidy_withingroup}
\alias{tidy_ingroup}
\title{Within-group effect by rebasing the group variable}
\usage{
tidy_withingroup(model, group, data, conf.int = TRUE, ..., .progress = FALSE)
}
\arguments{
\item{model}{The template model, will determine what method to use}

\item{group}{A vector of grouping variables}

\item{data}{Dataset for the performing subgroup. If missing, \code{model.frame(model)} will be inferred.
If the model object does not contains data, an error will be thrown.}

\item{conf.int}{[\code{TRUE}] include confidence interval}

\item{...}{additional parameters passed to the method}

\item{.progress}{[\code{FALSE}] Print progress?}
}
\value{
a tibble of class tidy_subgroup_tbl
}
\description{
This function rebase the group variable to different levels and refit the model, then tidy it.
The aim is to get the heterogeneous effect of other variables that have interaction with the group variable.
}
\details{
Different from \code{tidy_subgroup}, this function does not fit on the subcohort,
hence does not assume interaction between group variable and all other covariables in the model.
If one covariable does not have interaction with the group variable, the estimated effect would be unchanged.
}
\seealso{
\code{\link[generics:tidy]{generics::tidy()}} \code{\link[future:future]{future::future()}} \code{\link[future.apply:future_lapply]{future.apply::future_lapply()}} \code{\link[=tidy_subgroup]{tidy_subgroup()}}
}
\author{
Trinh Dong
}
