% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\name{forestploter}
\alias{forestploter}
\alias{forestploter.default}
\alias{forestploter.tidy_subgroup_tbl}
\title{Method to create forest plot for \code{tidy_subgroup} tibble}
\usage{
forestploter(x, ...)

\method{forestploter}{default}(
  data,
  est,
  lower,
  upper,
  sizes = 0.4,
  ref_line = ifelse(x_trans \%in\% c("log", "log2", "log10"), 1, 0),
  vert_line = NULL,
  ci_column,
  is_summary = NULL,
  xlim = NULL,
  ticks_at = NULL,
  ticks_digits = NULL,
  ticks_minor = NULL,
  arrow_lab = NULL,
  x_trans = "none",
  xlab = NULL,
  footnote = NULL,
  title = NULL,
  nudge_y = 0,
  fn_ci = makeci,
  fn_summary = make_summary,
  index_args = NULL,
  theme = NULL,
  ...
)

\method{forestploter}{tidy_subgroup_tbl}(
  x,
  terms = x$term[[2]],
  label = "Subgroup",
  stats = c(`Est. (95\% CI)` = format.ci(estimate, conf.low, conf.high, digits = 1),
    `p-value` = format.pval(p.value, digits = 2, eps = 0.001)),
  plot_pos = 2L,
  plot_width = 1,
  arrow_lab = c("Harmful", "Benefit"),
  ...
)
}
\arguments{
\item{x}{a tibble create by tidy_subgroup}

\item{...}{additional parameters passed to [forestploter::forest()].}

\item{arrow_lab}{[\code{c('Harmful', 'Benefit')}] label of the arrow}

\item{terms}{<character> which term to show forest plot}

\item{label}{[\code{"Subgroup"}] A character specifying the label of the subgroup column}

\item{stats}{stat names in x to be shown in the forest plot after the CI bar}

\item{plot_pos}{[\code{2}] <int> column position of plot}

\item{plot_width}{[\code{1}] <numeric> adjust plot width by a factor > 0}
}
\value{
a plot
}
\description{
This function add a method to create forest plot to tidied subgroup models.
}
\author{
Trinh Dong
}
