% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\name{forestploter}
\alias{forestploter}
\alias{forestploter.default}
\alias{forestploter.tidy_subgroup_tbl}
\alias{as_forest_dt}
\alias{as_forest_dt.defautl}
\alias{as_forest_dt.tidy_subgroup_tbl}
\title{Method to create forest plot for \code{tidy_subgroup} tibble}
\usage{
forestploter(x, ...)

\method{forestploter}{default}(x, stats_cols, ci_column, ...)

\method{forestploter}{tidy_subgroup_tbl}(
  x,
  terms = x$term[[2]],
  label = "Subgroup",
  stats = c(`Est. (95\% CI)` = format.ci(estimate, conf.low, conf.high, digits = 1),
    `p-value` = format.pval(p.value, digits = 2, eps = 0.001)),
  plot_pos = 2L,
  plot_width = 1,
  arrow_lab = c("Harmful", "Benefit"),
  .hide_label = FALSE,
  ...
)

as_forest_dt(x, ...)

\method{as_forest_dt}{defautl}(x, ...)

\method{as_forest_dt}{tidy_subgroup_tbl}(
  x,
  terms = x$term[[2]],
  label = "Subgroup",
  stats = c(`Est. (95\% CI)` = format.ci(estimate, conf.low, conf.high, digits = 1),
    `p-value` = format.pval(p.value, digits = 2, eps = 0.001)),
  plot_pos = 2L,
  plot_width = 1
)
}
\arguments{
\item{x}{a tibble create by tidy_subgroup}

\item{...}{additional parameters passed to [forestploter::forest()].}

\item{terms}{<character> which term to show forest plot}

\item{label}{[\code{"Subgroup"}] A character specifying the label of the subgroup column}

\item{stats}{stat names in x to be shown in the forest plot after the CI bar}

\item{plot_pos}{[\code{2}] <int> column position of plot}

\item{plot_width}{[\code{1}] <numeric> adjust plot width by a factor > 0}

\item{arrow_lab}{[\code{c('Harmful', 'Benefit')}] label of the arrow}

\item{.hide_label}{[\code{FALSE}] hide subgroup column.}
}
\value{
a plot

A tibble
}
\description{
This function add a method to create forest plot to tidied subgroup models.
}
\author{
Trinh Dong
}
