% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_subgroup.R
\name{tidy_subgroup}
\alias{tidy_subgroup}
\alias{as_tidy_subgroup}
\alias{as_tidy_subgroup.data.frame}
\title{Refit model on sub-cohort}
\usage{
tidy_subgroup(
  base_model,
  group,
  data,
  conf.int = TRUE,
  ...,
  .overall = FALSE,
  .progress = FALSE
)

as_tidy_subgroup(x, ...)

\method{as_tidy_subgroup}{data.frame}(x)
}
\arguments{
\item{base_model}{The main model, will determine what method to use}

\item{group}{A vector of grouping variables}

\item{data}{Dataset for the performing subgroup. If missing, \code{model.frame(base_model)} will be inferred.
If the model object does not contains data, an error will be thrown.}

\item{conf.int}{\link{\code{TRUE}} include confidence interval}

\item{...}{additional parameters passed to the method}

\item{.overall}{\link{\code{FALSE}} Save the overall model in the output?}

\item{.progress}{\link{\code{FALSE}} Print progress?}
}
\value{
a tibble of class tidy_subgroup_tbl
}
\description{
This function refits the model on sub-cohort, then tidy it.
}
\details{
This function relies on \code{tidy} methods. Therefore, as long as there is a suitable method for tidying the fit, this function should work.
}
\seealso{
\code{\link[generics:tidy]{generics::tidy()}} \code{\link[future:future]{future::future()}} \code{\link[future.apply:future_lapply]{future.apply::future_lapply()}} \code{\link[=tidy_ingroup]{tidy_ingroup()}}
}
\author{
Trinh Dong
}
