% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_subgroup.R
\name{tidy_subgroup}
\alias{tidy_subgroup}
\title{Refit model on sub-cohort}
\usage{
tidy_subgroup(
  base_model,
  group,
  data,
  conf.int = TRUE,
  ...,
  .overall = FALSE,
  .progress = FALSE
)
}
\arguments{
\item{base_model}{The main model, will determine that method to use}

\item{group}{A vector of grouping variables}

\item{data}{Dataset for the performing subgroup. If missing, \code{base_model$data} will be inferred.
If the model object does not contains data, an error will be thrown.}

\item{conf.int}{[\code{TRUE}] include confidence interval}

\item{...}{additional parameters passed to the method}

\item{.overall}{[\code{FALSE}] Save the overall model in the output?}

\item{.progress}{[\code{FALSE}] Print progress?}
}
\value{
a tibble
}
\description{
This function refits the model on sub-cohort, then tidy it.
}
\details{
This function relies on \code{tidy} methods. Therefore, as long as there is a suitable method for tidying the fit, this function should work.
}
\seealso{
[generics::tidy()] [future::future()] [future.apply::future_lapply()]
}
\author{
Trinh Dong
}
